+++ src/main/java/jwp/controller/CreateUserController.java
@@ -1,7 +1,6 @@
 package jwp.controller;
 
-import core.db.DBUserRepository;
-import core.db.MemoryUserRepository;
+import core.db.user.DBUserRepository;
 import core.mvc.Controller;
 import jwp.model.User;
