+++ src/main/java/jwp/controller/question/DeleteQuestionController.java
@@ -1,9 +1,8 @@
 package jwp.controller.question;
 
-import core.db.DBQuestionRepository;
-import core.db.DBUserRepository;
+import core.db.question.DBQuestionRepository;
+import core.db.user.DBUserRepository;
 import core.mvc.Controller;
-import jwp.model.Question;
 import jwp.model.User;
 
 import javax.servlet.http.HttpServletRequest;
